fix Firefox not prompting user/pass for CORS
Without withCredentials set, Firefox won't prompt the user for a
username and password for a cross origin request, even if the server has
all the proper CORS headers. With this change it is now possible to load
a kdbx file from another domain and be prompted for the username and
password.