Move Google Drive handling to a service.
Along the way, authorize to Google service in `app.js` too.
This is done so that we could determine whether we should use immediate in
the button click handler.

Previously, we try immediate=true first, then immediate=false. The problem
is immediate=true uses async XHR and by the time it completes, the button
click is no longer tied to it. So when we launch immediate=false, the pop
up blocker would jump in, because that action is not user-initiated.

So we silently try to authorize first. Then we obtain gapi.auth.getToken().
If there is not a token, we use immediate=false.

In the rare case that there is a token, and that token fails immediate=true,
then the popup blocker would still block us when we retry with immediate=false.
In that case, a simple refresh should be all it takes to fix.