Pull in gapi-chrome-apps lib.
This commit starts a new branch to unify Chrome app and regular web app
access to Google services.

The main problem with Chrome app is content security policy. Google API
libraries are loaded dynamically. And that is forbidden in CSP mode. So
we need to limit ourselves to REST API instead. That means the picker
must be reimplemented and we need to talk to Google Drive via `$http`.